Moves 'utils' into lib folder.
